requirements.txt
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.37
python-telegram-bot==21.3
ta==0.10.2  # 🚨 Alternative to ta-lib, pure Python, no C dependencies
schedule==1.2.2
requests==2.32.3
asyncio-mqtt==0.16.2
